// src/types/libraryTab.ts
export enum LibraryTab {
    Albums = 'albums',
    AlbumArtists = 'albumartists',
    Artists = 'artists',
    Channels = 'channels',
    Collections = 'collections',
    Episodes = 'episodes',
    Favorites = 'favorites',
    Genres = 'genres',
    Guide = 'guide',
    Movies = 'movies',
    Networks = 'networks',
    Playlists = 'playlists',
    Programs = 'programs',
    Recordings = 'recordings',
    Schedule = 'schedule',
    Series = 'series',
    SeriesTimers = 'seriestimers',
    Songs = 'songs',
    Suggestions = 'suggestions',
    Upcoming = 'upcoming',
    PhotoAlbums = 'photoalbums',
    Photos = 'photos',
    Videos = 'videos',
    Books = 'books',
    Years = 'years'
}